Merged PR 194487: Changed announceForAccessibility implementation to use Assertive live region
Changed announceForAccessibility implementation to use Assertive live region type instead of Polite. When Polite is used the request for announcement will be overridden by a subsequent narration. Usually this happens when code calls the announcement API and then focus is changed either programmatically or due to UI changes as result of event that was announced - in this case the automatic narration of a newly focused control suppresses the announcement, usually even before it starts. With announcement being Assertive, though, both automatic narration and the announcement are read and no information is lost.

Related work items: #1495361