Added support for multi window apps to Alert.
A committed but not yet released version of RN for Windows adds support for alerts in multi window apps by the means of a new option in RN.AlertOptions. This option is a react tag that is used as a hint by the native side to detect the right top level window to parent the native alert.
ReactXP API models the identification of root views through "rootViewId" properties. They are already used for modals/popups, so we're adding same property to RXP.AlertOptions.
The implementation keeps a list of active root views in order to quickly map the rootViewId to the RNW expected react tag hint.
This mechanism will be used for other APIs as well in future PRs (there's one where we pass the rootViewId to RNW directly, and that in not very clean since it expects RNW to understand RXP properties)

(cherry picked before merging to master: https://github.com/Microsoft/reactxp/pull/787)